+++ graphile/graphile-settings/src/presets/constructive-preset.ts
@@ -11,7 +11,6 @@ import {
   MetaSchemaPreset,
   PgTypeMappingsPreset,
   RequiredInputPreset,
-  NodeTypeRegistryPreset,
 } from '../plugins';
 import { UnifiedSearchPreset, createMatchesOperatorFactory, createTrgmOperatorFactories } from 'graphile-search';
 import { GraphilePostgisPreset, createPostgisOperatorFactory } from 'graphile-postgis';
@@ -91,7 +90,6 @@ export const ConstructivePreset: GraphileConfig.Preset = {
     SqlExpressionValidatorPreset(),
     PgTypeMappingsPreset,
     RequiredInputPreset,
-    NodeTypeRegistryPreset,
   ],
   /**
    * Disable PostGraphile core's condition argument entirely.