+++ graphile/node-type-registry/src/relation/index.ts
@@ -1,0 +1,4 @@
+export { RelationBelongsTo } from './relation-belongs-to';
+export { RelationHasOne } from './relation-has-one';
+export { RelationHasMany } from './relation-has-many';
+export { RelationManyToMany } from './relation-many-to-many';