+++ graphile/node-type-registry/src/table/index.ts
@@ -1,0 +1,3 @@
+export { TableUserProfiles } from './table-user-profiles';
+export { TableOrganizationSettings } from './table-organization-settings';
+export { TableUserSettings } from './table-user-settings';