+++ graphile/graphile-postgis/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.3](https://github.com/constructive-io/constructive/compare/graphile-postgis@2.8.2...graphile-postgis@2.8.3) (2026-03-25)
+
+**Note:** Version bump only for package graphile-postgis
+
+## [2.8.2](https://github.com/constructive-io/constructive/compare/graphile-postgis@2.8.1...graphile-postgis@2.8.2) (2026-03-25)
+
+**Note:** Version bump only for package graphile-postgis
+
+## [2.8.1](https://github.com/constructive-io/constructive/compare/graphile-postgis@2.8.0...graphile-postgis@2.8.1) (2026-03-24)
+
+**Note:** Version bump only for package graphile-postgis
+
+# [2.8.0](https://github.com/constructive-io/constructive/compare/graphile-postgis@2.7.2...graphile-postgis@2.8.0) (2026-03-24)
+
+**Note:** Version bump only for package graphile-postgis
+
 ## [2.7.2](https://github.com/constructive-io/constructive/compare/graphile-postgis@2.7.1...graphile-postgis@2.7.2) (2026-03-21)
 
 **Note:** Version bump only for package graphile-postgis