+++ graphile/graphile-test/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.6.3](https://github.com/constructive-io/constructive/compare/graphile-test@4.6.2...graphile-test@4.6.3) (2026-03-25)
+
+**Note:** Version bump only for package graphile-test
+
+## [4.6.2](https://github.com/constructive-io/constructive/compare/graphile-test@4.6.1...graphile-test@4.6.2) (2026-03-25)
+
+**Note:** Version bump only for package graphile-test
+
+## [4.6.1](https://github.com/constructive-io/constructive/compare/graphile-test@4.6.0...graphile-test@4.6.1) (2026-03-24)
+
+**Note:** Version bump only for package graphile-test
+
+# [4.6.0](https://github.com/constructive-io/constructive/compare/graphile-test@4.5.10...graphile-test@4.6.0) (2026-03-24)
+
+**Note:** Version bump only for package graphile-test
+
 ## [4.5.10](https://github.com/constructive-io/constructive/compare/graphile-test@4.5.9...graphile-test@4.5.10) (2026-03-21)
 
 **Note:** Version bump only for package graphile-test