+++ graphql/orm-test/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.6](https://github.com/constructive-io/constructive/compare/graphql-orm-test@0.3.5...graphql-orm-test@0.3.6) (2026-03-25)
+
+**Note:** Version bump only for package graphql-orm-test
+
+## [0.3.5](https://github.com/constructive-io/constructive/compare/graphql-orm-test@0.3.4...graphql-orm-test@0.3.5) (2026-03-25)
+
+**Note:** Version bump only for package graphql-orm-test
+
+## [0.3.4](https://github.com/constructive-io/constructive/compare/graphql-orm-test@0.3.3...graphql-orm-test@0.3.4) (2026-03-25)
+
+**Note:** Version bump only for package graphql-orm-test
+
+## [0.3.3](https://github.com/constructive-io/constructive/compare/graphql-orm-test@0.3.2...graphql-orm-test@0.3.3) (2026-03-25)
+
+**Note:** Version bump only for package graphql-orm-test
+
+## [0.3.2](https://github.com/constructive-io/constructive/compare/graphql-orm-test@0.3.1...graphql-orm-test@0.3.2) (2026-03-25)
+
+**Note:** Version bump only for package graphql-orm-test
+
+## [0.3.1](https://github.com/constructive-io/constructive/compare/graphql-orm-test@0.3.0...graphql-orm-test@0.3.1) (2026-03-24)
+
+**Note:** Version bump only for package graphql-orm-test
+
+# [0.3.0](https://github.com/constructive-io/constructive/compare/graphql-orm-test@0.2.2...graphql-orm-test@0.3.0) (2026-03-24)
+
+**Note:** Version bump only for package graphql-orm-test
+
 ## [0.2.2](https://github.com/constructive-io/constructive/compare/graphql-orm-test@0.2.1...graphql-orm-test@0.2.2) (2026-03-21)
 
 **Note:** Version bump only for package graphql-orm-test