+++ graphql/query/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.10.4](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-query@3.10.3...@constructive-io/graphql-query@3.10.4) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/graphql-query
+
+## [3.10.3](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-query@3.10.2...@constructive-io/graphql-query@3.10.3) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/graphql-query
+
+## [3.10.2](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-query@3.10.1...@constructive-io/graphql-query@3.10.2) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/graphql-query
+
+## [3.10.1](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-query@3.10.0...@constructive-io/graphql-query@3.10.1) (2026-03-24)
+
+**Note:** Version bump only for package @constructive-io/graphql-query
+
+# [3.10.0](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-query@3.9.1...@constructive-io/graphql-query@3.10.0) (2026-03-24)
+
+**Note:** Version bump only for package @constructive-io/graphql-query
+
 ## [3.9.1](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-query@3.9.0...@constructive-io/graphql-query@3.9.1) (2026-03-21)
 
 **Note:** Version bump only for package @constructive-io/graphql-query