+++ graphql/server/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.14.4](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-server@4.14.3...@constructive-io/graphql-server@4.14.4) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/graphql-server
+
+## [4.14.3](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-server@4.14.2...@constructive-io/graphql-server@4.14.3) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/graphql-server
+
+## [4.14.2](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-server@4.14.1...@constructive-io/graphql-server@4.14.2) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/graphql-server
+
+## [4.14.1](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-server@4.14.0...@constructive-io/graphql-server@4.14.1) (2026-03-24)
+
+**Note:** Version bump only for package @constructive-io/graphql-server
+
+# [4.14.0](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-server@4.13.2...@constructive-io/graphql-server@4.14.0) (2026-03-24)
+
+**Note:** Version bump only for package @constructive-io/graphql-server
+
 ## [4.13.2](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-server@4.13.1...@constructive-io/graphql-server@4.13.2) (2026-03-21)
 
 **Note:** Version bump only for package @constructive-io/graphql-server