+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.10.6](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@7.10.5...@constructive-io/cli@7.10.6) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [7.10.5](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@7.10.4...@constructive-io/cli@7.10.5) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [7.10.4](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@7.10.3...@constructive-io/cli@7.10.4) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [7.10.3](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@7.10.2...@constructive-io/cli@7.10.3) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [7.10.2](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@7.10.1...@constructive-io/cli@7.10.2) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [7.10.1](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@7.10.0...@constructive-io/cli@7.10.1) (2026-03-24)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+# [7.10.0](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@7.9.3...@constructive-io/cli@7.10.0) (2026-03-24)
+
+**Note:** Version bump only for package @constructive-io/cli
+
 ## [7.9.3](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@7.9.2...@constructive-io/cli@7.9.3) (2026-03-21)
 
 **Note:** Version bump only for package @constructive-io/cli