+++ pgpm/core/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.9.3](https://github.com/constructive-io/constructive/compare/@pgpmjs/core@6.9.2...@pgpmjs/core@6.9.3) (2026-03-25)
+
+**Note:** Version bump only for package @pgpmjs/core
+
+## [6.9.2](https://github.com/constructive-io/constructive/compare/@pgpmjs/core@6.9.1...@pgpmjs/core@6.9.2) (2026-03-25)
+
+**Note:** Version bump only for package @pgpmjs/core
+
+## [6.9.1](https://github.com/constructive-io/constructive/compare/@pgpmjs/core@6.9.0...@pgpmjs/core@6.9.1) (2026-03-24)
+
+**Note:** Version bump only for package @pgpmjs/core
+
+# [6.9.0](https://github.com/constructive-io/constructive/compare/@pgpmjs/core@6.8.2...@pgpmjs/core@6.9.0) (2026-03-24)
+
+### Bug Fixes
+
+- **export:** use extensionName for schema prefix in meta replacer (parity with SQL flow) ([681d3b9](https://github.com/constructive-io/constructive/commit/681d3b9a89cabfb2a3771f306477828232f0ddf6))
+
 ## [6.8.2](https://github.com/constructive-io/constructive/compare/@pgpmjs/core@6.8.1...@pgpmjs/core@6.8.2) (2026-03-21)
 
 **Note:** Version bump only for package @pgpmjs/core