+++ postgres/introspectron/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.8.4](https://github.com/constructive-io/constructive/compare/introspectron@4.8.3...introspectron@4.8.4) (2026-03-25)
+
+**Note:** Version bump only for package introspectron
+
+## [4.8.3](https://github.com/constructive-io/constructive/compare/introspectron@4.8.2...introspectron@4.8.3) (2026-03-25)
+
+**Note:** Version bump only for package introspectron
+
+## [4.8.2](https://github.com/constructive-io/constructive/compare/introspectron@4.8.1...introspectron@4.8.2) (2026-03-25)
+
+**Note:** Version bump only for package introspectron
+
+## [4.8.1](https://github.com/constructive-io/constructive/compare/introspectron@4.8.0...introspectron@4.8.1) (2026-03-24)
+
+**Note:** Version bump only for package introspectron
+
+# [4.8.0](https://github.com/constructive-io/constructive/compare/introspectron@4.7.2...introspectron@4.8.0) (2026-03-24)
+
+**Note:** Version bump only for package introspectron
+
 ## [4.7.2](https://github.com/constructive-io/constructive/compare/introspectron@4.7.1...introspectron@4.7.2) (2026-03-21)
 
 **Note:** Version bump only for package introspectron