+++ sdk/constructive-node/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.6](https://github.com/constructive-io/constructive/compare/@constructive-io/node@0.7.5...@constructive-io/node@0.7.6) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/node
+
+## [0.7.5](https://github.com/constructive-io/constructive/compare/@constructive-io/node@0.7.4...@constructive-io/node@0.7.5) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/node
+
+## [0.7.4](https://github.com/constructive-io/constructive/compare/@constructive-io/node@0.7.3...@constructive-io/node@0.7.4) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/node
+
+## [0.7.3](https://github.com/constructive-io/constructive/compare/@constructive-io/node@0.7.2...@constructive-io/node@0.7.3) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/node
+
+## [0.7.2](https://github.com/constructive-io/constructive/compare/@constructive-io/node@0.7.1...@constructive-io/node@0.7.2) (2026-03-25)
+
+**Note:** Version bump only for package @constructive-io/node
+
+## [0.7.1](https://github.com/constructive-io/constructive/compare/@constructive-io/node@0.7.0...@constructive-io/node@0.7.1) (2026-03-24)
+
+**Note:** Version bump only for package @constructive-io/node
+
+# [0.7.0](https://github.com/constructive-io/constructive/compare/@constructive-io/node@0.6.5...@constructive-io/node@0.7.0) (2026-03-24)
+
+**Note:** Version bump only for package @constructive-io/node
+
 ## [0.6.5](https://github.com/constructive-io/constructive/compare/@constructive-io/node@0.6.4...@constructive-io/node@0.6.5) (2026-03-22)
 
 **Note:** Version bump only for package @constructive-io/node